+++ src/Exception/RequestUriTooLongException.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 namespace Narrowspark\HttpStatus\Exception;
 
 class RequestUriTooLongException extends AbstractClientErrorException
@@ -6,7 +7,7 @@ class RequestUriTooLongException extends AbstractClientErrorException
     /**
      * @var string
      */
-    protected $message = '414 Request-URI Too Long';
+    protected $message = '414 URI Too Long';
 
     /**
      * @var int