+++ src/Exception/RequestedRangeNotSatisfiableException.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 namespace Narrowspark\HttpStatus\Exception;
 
 class RequestedRangeNotSatisfiableException extends AbstractClientErrorException
@@ -6,7 +7,7 @@ class RequestedRangeNotSatisfiableException extends AbstractClientErrorException
     /**
      * @var string
      */
-    protected $message = '416 Requested Range Not Satisfiable';
+    protected $message = '416 Range Not Satisfiable';
 
     /**
      * @var int