+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Changelog
 
 All Notable changes to `http-status` will be documented in this file
+## NEXT - 2017-02-20
+
+### Added
+- MisdirectedRequestException, the error code 421 and the status message.
+- Support for [http-message-util](//github.com/php-fig/http-message-util)
+- Tests against iana.org
+
+### Deprecated
+- Nothing
+
+### Fixed
+- Nothing
+
+### Removed
+- Support for php 5.6 and hhvm
+
+### Security
+- Nothing
 
 ## NEXT - 2016-06-22